Move 'tour' to core/commands
Side effect: this makes 'tour' accessible through the HTTP API
