+++ cmd/ipfs/ipfs.go
@@ -23,7 +23,6 @@ var commandsClientCmd = commands.CommandsCmd(Root)
 var localCommands = map[string]*cmds.Command{
 	"daemon":   daemonCmd,
 	"init":     initCmd,
-	"tour":     tourCmd,
 	"commands": commandsClientCmd,
 }
 var localMap = make(map[*cmds.Command]bool)
